+++ src/com/gamesbykevin/puzzle/shapes/Hexagon.java
@@ -20,23 +20,23 @@ public final class Hexagon extends Piece implements PuzzlePiece
         int[] xpoints = new int[6];
         int[] ypoints = new int[6];
         
-        xpoints[0] = getX();
-        ypoints[0] = getY() - (getHeight() / 2);
+        xpoints[0] = (int)getX();
+        ypoints[0] = (int)(getY() - (getHeight() / 2));
         
         xpoints[1] = (int)(getX() + radius);
-        ypoints[1] = getY() - (getHeight() / 4);
+        ypoints[1] = (int)(getY() - (getHeight() / 4));
         
         xpoints[2] = (int)(getX() + radius);
-        ypoints[2] = getY() + (getHeight() / 4);
+        ypoints[2] = (int)(getY() + (getHeight() / 4));
         
-        xpoints[3] = getX();
-        ypoints[3] = getY() + (getHeight() / 2);
+        xpoints[3] = (int)getX();
+        ypoints[3] = (int)(getY() + (getHeight() / 2));
         
         xpoints[4] = (int)(getX() - radius);
-        ypoints[4] = getY() + (getHeight() / 4);
+        ypoints[4] = (int)(getY() + (getHeight() / 4));
         
         xpoints[5] = (int)(getX() - radius);
-        ypoints[5] = getY() - (getHeight() / 4);
+        ypoints[5] = (int)(getY() - (getHeight() / 4));
         
         Polygon p = new Polygon(xpoints, ypoints, xpoints.length);;
